Define default mounts for Laravel. (#75)
Define default mounts for Laravel.

Out of the box, Laravel uses /storage/app as the base local disk. It's a good default. Either way we need to explain to users what mounts are as this is a big part of migrating an app to us.

Our Laravel template defines mounts, we may rely on it: https://github.com/platformsh/template-builder/blob/master/templates/laravel/files/.platform.app.yaml#L57-L79

Closes https://github.com/platformsh/platformify/issues/67